requirements.txt
discord.py==2.3.1
emoji==2.8.0
python-dotenv==1.0.0
